fix pylint error
