+++ friends/utils.py
@@ -1,4 +1,5 @@
 import sys
+from itsdangerous import URLSafeSerializer
 
 
 def import_module(name):
@@ -14,3 +15,16 @@ def module_member(name):
 
 def get_strategy(strategy_cls, storage, *args, **kwargs):
     return strategy_cls(storage, *args, **kwargs)
+
+
+def get_serializer(strategy):
+    serializer = URLSafeSerializer(strategy.encryption_key())
+    return serializer
+
+
+def make_token(strategy, from_user, to_user):
+    payload = {
+        'from_user_id': "%s" % strategy.storage.user.get_id(from_user),
+        'to_user_id': "%s" % strategy.storage.user.get_id(to_user)
+    }
+    return get_serializer(strategy).dumps(payload).encode('utf-8')
